feat(js): export AuthRequest in js as a wrapper around native
