--- pubky/clippy.toml
@@ -1,1 +1,0 @@
-allow-unwrap-in-tests = true
