+++ pubky/src/native.rs
@@ -1,28 +1,49 @@
+pub mod internal {
+    #[cfg(not(wasm_browser))]
+    pub mod cookies;
+    pub mod pkarr;
+}
+pub mod api {
+    pub mod auth;
+    #[cfg(not(wasm_browser))]
+    pub mod http;
+    pub mod public;
+}
+
+use std::fmt::Debug;
+
+#[cfg(not(wasm_browser))]
 use std::{sync::Arc, time::Duration};
 
+#[cfg(not(wasm_browser))]
 use mainline::Testnet;
 
-use crate::Client;
+static DEFAULT_USER_AGENT: &str = concat!("pubky.org", "@", env!("CARGO_PKG_VERSION"),);
 
-mod api;
-mod cookies;
-mod http;
+#[macro_export]
+macro_rules! handle_http_error {
+    ($res:expr) => {
+        if let Err(status) = $res.error_for_status_ref() {
+            return match $res.text().await {
+                Ok(text) => Err(anyhow::anyhow!("{status}. Error message: {text}")),
+                _ => Err(anyhow::anyhow!("{status}")),
+            };
+        }
+    };
+}
 
-pub(crate) use cookies::CookieJar;
-
-static DEFAULT_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"),);
-
-#[derive(Debug, Default)]
+#[derive(Debug, Default, Clone)]
 pub struct ClientBuilder {
     pkarr: pkarr::ClientBuilder,
 }
 
 impl ClientBuilder {
+    #[cfg(not(wasm_browser))]
     /// Sets the following:
     /// - Pkarr client's DHT bootstrap nodes = `testnet` bootstrap nodes.
     /// - Pkarr client's resolvers           = `testnet` bootstrap nodes.
     /// - Pkarr client's DHT request timeout  = 500 milliseconds. (unless in CI, then it is left as default 2000)
-    pub fn testnet(mut self, testnet: &Testnet) -> Self {
+    pub fn testnet(&mut self, testnet: &Testnet) -> &mut Self {
         let bootstrap = testnet.bootstrap.clone();
 
         self.pkarr.bootstrap(&bootstrap);
@@ -34,15 +55,27 @@ impl ClientBuilder {
         self
     }
 
+    /// Create a [mainline::DhtBuilder] if `None`, and allows mutating it with a callback function.
+    pub fn pkarr<F>(&mut self, f: F) -> &mut Self
+    where
+        F: FnOnce(&mut pkarr::ClientBuilder) -> &mut pkarr::ClientBuilder,
+    {
+        f(&mut self.pkarr);
+
+        self
+    }
+
     /// Build [Client]
-    pub fn build(self) -> Result<Client, BuildError> {
+    pub fn build(&self) -> Result<Client, BuildError> {
         let pkarr = self.pkarr.build()?;
 
-        let cookie_store = Arc::new(CookieJar::default());
+        #[cfg(not(wasm_browser))]
+        let cookie_store = Arc::new(internal::cookies::CookieJar::default());
 
         // TODO: allow custom user agent, but force a Pubky user agent information
         let user_agent = DEFAULT_USER_AGENT;
 
+        #[cfg(not(wasm_browser))]
         let http = reqwest::ClientBuilder::from(pkarr.clone())
             // TODO: use persistent cookie jar
             .cookie_provider(cookie_store.clone())
@@ -50,19 +83,52 @@ impl ClientBuilder {
             .build()
             .expect("config expected to not error");
 
-        let icann_http = reqwest::ClientBuilder::new()
-            .cookie_provider(cookie_store.clone())
+        #[cfg(wasm_browser)]
+        let http = reqwest::Client::builder()
             .user_agent(user_agent)
             .build()
             .expect("config expected to not error");
 
         Ok(Client {
-            cookie_store,
             http,
-            icann_http,
             pkarr,
+
+            #[cfg(not(wasm_browser))]
+            icann_http: reqwest::Client::builder()
+                // TODO: use persistent cookie jar
+                .cookie_provider(cookie_store.clone())
+                .user_agent(user_agent)
+                .build()
+                .expect("config expected to not error"),
+            #[cfg(not(wasm_browser))]
+            cookie_store,
+
+            #[cfg(wasm_browser)]
+            testnet: false,
         })
     }
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum BuildError {
+    #[error(transparent)]
+    /// Error building Pkarr client.
+    PkarrBuildError(#[from] pkarr::errors::BuildError),
+}
+
+/// A client for Pubky homeserver API, as well as generic HTTP requests to Pubky urls.
+#[derive(Clone, Debug)]
+pub struct Client {
+    pub(crate) http: reqwest::Client,
+    pub(crate) pkarr: pkarr::Client,
+
+    #[cfg(not(wasm_browser))]
+    pub(crate) cookie_store: std::sync::Arc<internal::cookies::CookieJar>,
+    #[cfg(not(wasm_browser))]
+    pub(crate) icann_http: reqwest::Client,
+
+    #[cfg(wasm_browser)]
+    pub(crate) testnet: bool,
 }
 
 impl Client {
@@ -71,6 +137,7 @@ impl Client {
         ClientBuilder::default()
     }
 
+    #[cfg(not(wasm_browser))]
     /// Create a client connected to the local network
     /// with the bootstrapping node: `localhost:6881`
     pub fn testnet() -> Result<Self, BuildError> {
@@ -83,15 +150,9 @@ impl Client {
     }
 
     #[cfg(test)]
+    #[cfg(not(wasm_browser))]
     /// Alias to `pubky::Client::builder().testnet(testnet).build().unwrap()`
     pub(crate) fn test(testnet: &Testnet) -> Client {
         Client::builder().testnet(testnet).build().unwrap()
     }
-}
-
-#[derive(Debug, thiserror::Error)]
-pub enum BuildError {
-    #[error(transparent)]
-    /// Error building Pkarr client.
-    PkarrBuildError(#[from] pkarr::errors::BuildError),
 }
