+++ pubky/src/wasm.rs
@@ -1,10 +1,21 @@
 use wasm_bindgen::prelude::*;
 
-use crate::Client;
+pub mod api {
+    pub mod auth;
+    pub mod http;
+    pub mod public;
+    pub mod recovery_file;
+}
 
-mod api;
-mod http;
-mod wrappers;
+pub mod wrappers {
+    pub mod keys;
+    pub mod session;
+}
+
+static TESTNET_RELAYS: [&str; 1] = ["http://localhost:15411/"];
+
+#[wasm_bindgen]
+pub struct Client(crate::NativeClient);
 
 impl Default for Client {
     fn default() -> Self {
@@ -12,18 +23,16 @@ impl Default for Client {
     }
 }
 
-static TESTNET_RELAYS: [&str; 1] = ["http://localhost:15411/"];
-
 #[wasm_bindgen]
 impl Client {
     #[wasm_bindgen(constructor)]
     /// Create Client with default Settings including default relays
     pub fn new() -> Self {
-        Self {
-            http: reqwest::Client::builder().build().unwrap(),
-            pkarr: pkarr::Client::builder().build().unwrap(),
-            testnet: false,
-        }
+        Self(
+            crate::NativeClient::builder()
+                .build()
+                .expect("building a default NativeClient should be infallible"),
+        )
     }
 
     /// Create a client with with configurations appropriate for local testing:
@@ -32,15 +41,19 @@ impl Client {
     ///     and read the homeserver HTTP port from the [reserved service parameter key](pubky_common::constants::reserved_param_keys::HTTP_PORT)
     #[wasm_bindgen]
     pub fn testnet() -> Self {
-        Self {
-            http: reqwest::Client::builder().build().unwrap(),
-            pkarr: pkarr::Client::builder()
+        let mut builder = crate::NativeClient::builder();
+
+        builder.pkarr(|builder| {
+            builder
                 .relays(&TESTNET_RELAYS)
                 .expect("testnet relays are valid urls")
-                .build()
-                .unwrap(),
-            testnet: true,
-        }
+        });
+
+        let mut client = builder.build().expect("testnet build should be infallibl");
+
+        client.testnet = true;
+
+        Self(client)
     }
 }
