+++ pubky/src/wasm/api/auth.rs
@@ -4,10 +4,12 @@ use url::Url;
 
 use pubky_common::capabilities::Capabilities;
 
-use crate::Client;
+use crate::wasm::wrappers::{
+    keys::{Keypair, PublicKey},
+    session::Session,
+};
 
-use crate::wasm::wrappers::keys::{Keypair, PublicKey};
-use crate::wasm::wrappers::session::Session;
+use super::super::Client;
 
 use wasm_bindgen::prelude::*;
 
@@ -24,19 +26,21 @@ impl Client {
         homeserver: &PublicKey,
     ) -> Result<Session, JsValue> {
         Ok(Session(
-            self.inner_signup(keypair.as_inner(), homeserver.as_inner())
+            self.0
+                .signup(keypair.as_inner(), homeserver.as_inner())
                 .await
                 .map_err(|e| JsValue::from_str(&e.to_string()))?,
         ))
     }
 
-    /// Check the current sesison for a given Pubky in its homeserver.
+    /// Check the current session for a given Pubky in its homeserver.
     ///
-    /// Returns [Session] or `None` (if recieved `404 NOT_FOUND`),
-    /// or throws the recieved error if the response has any other `>=400` status code.
+    /// Returns [Session] or `None` (if received `404 NOT_FOUND`),
+    /// or throws the received error if the response has any other `>=400` status code.
     #[wasm_bindgen]
     pub async fn session(&self, pubky: &PublicKey) -> Result<Option<Session>, JsValue> {
-        self.inner_session(pubky.as_inner())
+        self.0
+            .session(pubky.as_inner())
             .await
             .map(|s| s.map(Session))
             .map_err(|e| JsValue::from_str(&e.to_string()))
@@ -45,7 +49,8 @@ impl Client {
     /// Signout from a homeserver.
     #[wasm_bindgen]
     pub async fn signout(&self, pubky: &PublicKey) -> Result<(), JsValue> {
-        self.inner_signout(pubky.as_inner())
+        self.0
+            .signout(pubky.as_inner())
             .await
             .map_err(|e| JsValue::from_str(&e.to_string()))
     }
@@ -53,7 +58,8 @@ impl Client {
     /// Signin to a homeserver using the root Keypair.
     #[wasm_bindgen]
     pub async fn signin(&self, keypair: &Keypair) -> Result<(), JsValue> {
-        self.inner_signin(keypair.as_inner())
+        self.0
+            .signin(keypair.as_inner())
             .await
             .map(|_| ())
             .map_err(|e| JsValue::from_str(&e.to_string()))
@@ -63,35 +69,18 @@ impl Client {
     /// verifying that AuthToken, and if capabilities were requested, signing in to
     /// the Pubky's homeserver and returning the [Session] information.
     ///
-    /// Returns a tuple of [pubkyAuthUrl, Promise<Session>]
+    /// Returns a [AuthRequest]
     #[wasm_bindgen(js_name = "authRequest")]
-    pub fn auth_request(&self, relay: &str, capabilities: &str) -> Result<js_sys::Array, JsValue> {
-        let mut relay: Url = relay.try_into().map_err(|_| "Invalid relay Url")?;
-
-        let (pubkyauth_url, client_secret) = self
-            .create_auth_request(
-                &mut relay,
+    pub fn auth_request(&self, relay: &str, capabilities: &str) -> Result<AuthRequest, JsValue> {
+        let auth_request = self
+            .0
+            .auth_request(
+                relay,
                 &Capabilities::try_from(capabilities).map_err(|_| "Invalid capaiblities")?,
             )
             .map_err(|e| JsValue::from_str(&e.to_string()))?;
 
-        let this = self.clone();
-
-        let future = async move {
-            this.subscribe_to_auth_response(relay, &client_secret)
-                .await
-                .map(|pubky| JsValue::from(PublicKey(pubky)))
-                .map_err(|e| JsValue::from_str(&e.to_string()))
-        };
-
-        let promise = wasm_bindgen_futures::future_to_promise(future);
-
-        // Return the URL and the promise
-        let js_tuple = js_sys::Array::new();
-        js_tuple.push(&JsValue::from_str(pubkyauth_url.as_ref()));
-        js_tuple.push(&promise);
-
-        Ok(js_tuple)
+        Ok(AuthRequest(auth_request))
     }
 
     /// Sign an [pubky_common::auth::AuthToken], encrypt it and send it to the
@@ -104,10 +93,40 @@ impl Client {
     ) -> Result<(), JsValue> {
         let pubkyauth_url: Url = pubkyauth_url.try_into().map_err(|_| "Invalid relay Url")?;
 
-        self.inner_send_auth_token(keypair.as_inner(), pubkyauth_url)
+        self.0
+            .send_auth_token(keypair.as_inner(), &pubkyauth_url)
             .await
             .map_err(|e| JsValue::from_str(&e.to_string()))?;
 
         Ok(())
+    }
+}
+
+#[wasm_bindgen]
+pub struct AuthRequest(crate::AuthRequest);
+
+#[wasm_bindgen]
+impl AuthRequest {
+    /// Returns the Pubky Auth url, which you should show to the user
+    /// to request an authentication or authorization token.
+    ///
+    /// Wait for this token using `this.response()`.
+    #[wasm_bindgen]
+    pub fn url(&self) -> String {
+        self.0.url().as_str().to_string()
+    }
+
+    /// Wait for the user to send an authentication or authorization proof.
+    ///
+    /// If successful, you should expect an instance of [PublicKey]
+    ///
+    /// Otherwise it will throw an error.
+    #[wasm_bindgen]
+    pub async fn response(&self) -> Result<PublicKey, JsValue> {
+        self.0
+            .response()
+            .await
+            .map(PublicKey::from)
+            .map_err(|e| JsValue::from_str(&e.to_string()))
     }
 }
