--- pubky/src/wasm/api/mod.rs
@@ -1,5 +1,0 @@
-pub mod recovery_file;
-
-// TODO: put the Homeserver API behind a feature flag
-pub mod auth;
-pub mod public;
