--- pubky/src/wasm/wrappers/mod.rs
@@ -1,5 +1,0 @@
-//! Wasm wrappers around structs that we need to be turned into Classes
-//! in JavaScript.
-
-pub mod keys;
-pub mod session;
